server/requirements.txt
boto3==1.26.99
Flask==2.2.2
flask_sqlalchemy==3.0.3
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0